requirements.txt
aiomysql==0.2.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
bcrypt==4.3.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
cliapp==1.20180812.1
click==8.1.8
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==44.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
duplicity==0.8.21
ecdsa==0.19.1
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
fastapi==0.115.12
fasteners==0.14.1
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.4
future==0.18.2
greenlet==3.2.1
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.1.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.10
Markdown==3.3.6
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==25.0
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.5.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-unixsocket==0.2.0
rsa==4.9.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
systemd-python==234
tomli==2.2.1
ttystatus==0.38
typing-inspection==0.4.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.2
wadllib==1.3.6
Werkzeug==3.0.4
xdg==5
xkit==0.0.0
zipp==1.0.0
